Twilio SMS

+++ backend/src/main/java/com/tiago/dsmeta/repositories/SaleRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.tiago.dsmeta.entities.Sale;
 
-public interface SaleRepository extends JpaRepository<Sale, Integer> {
+public interface SaleRepository extends JpaRepository<Sale, Long> {
 
 	@Query("SELECT obj FROM Sale obj WHERE obj.date BETWEEN :min AND :max ORDER BY obj.amount DESC")
 	Page<Sale> findSales(LocalDate min, LocalDate max, Pageable pageable);
